+++ loggers/winston/index.js
@@ -79,7 +79,7 @@ function ecsTransform (info, opts) {
   }
 
   // If there is a *started* APM agent, then use it.
-  const apm = elasticApm && elasticApm.isStarted() ? elasticApm : null
+  const apm = elasticApm && elasticApm.isStarted && elasticApm.isStarted() ? elasticApm : null
 
   // istanbul ignore else
   if (apm) {