+++ loggers/pino/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- Add support for [ECS tracing fields](https://www.elastic.co/guide/en/ecs/current/ecs-tracing.html).
+  If it is detected that [Elastic APM](https://www.npmjs.com/package/elastic-apm-node)
+  is in use and there is an active trace, then tracing fields will be added to
+  log records. This enables linking between traces and log records in Kibana.
+
 - BREAKING CHANGE: Conversion of HTTP request and response objects is no longer
   done by default. One must use the new `convertReqRes: true` formatter option.
   As well, only the meta keys `req` and `res` will be handled. Before this